+++ components/friend/search/SearchInput.tsx
