+++ components/friend/search/SearchTitle.tsx
